refactor(device-discovery): extract switch-stack translation into translate_chassis.py
translate.py was at 968 lines after the stack/VC additions. The stack-emission
code is a self-contained ~290-line cluster that only fires when chassis_members
is present, so split it out:

  - device_discovery/translate_chassis.py (new, 335 lines):
    validate_chassis_payload, translate_as_stack, _master_device_ref,
    _route_interfaces_by_member, and per-member device/interface builders
  - device_discovery/translate.py (677 lines, down from 968):
    keeps single-Device path, calls into translate_chassis on the stack branch

Public surface from translate_chassis: validate_chassis_payload and
translate_as_stack (renamed without leading underscore since they cross the
module boundary). translate.py imports both at the top; translate_chassis
imports translate_device and assign_primary_ip from translate.py via local
function-scoped imports to avoid circular module-load.

Tests + ruff unchanged. Sets up a clean seam for batch-2 vendor translators
(Junos VC, NX-OS VPC) to live alongside Cisco StackWise without bloating
translate.py further.
